Fix Application export
