+++ pkg/epinio/models/applications.js
@@ -1,7 +1,6 @@
 import { classify } from '@shell/plugins/dashboard-store/classify';
 import { downloadFile } from '@shell/utils/download';
 import { formatSi } from '@shell/utils/units';
-import JSZip from 'jszip';
 import { identity, pickBy } from 'lodash';
 import { epiniofy } from '../store/epinio-store/actions';
 import {
@@ -591,46 +590,6 @@ export default class EpinioApplicationModel extends EpinioNamespacedResource {
       await downloadFile(`${ this.meta.name }-${ part }.yaml`, data, 'text/plain');
     } else {
       await downloadFile(`${ this.meta.name }-${ part }`, data, 'application/gzip;charset=utf-8');
-    }
-  }
-
-  // TODO: Remove after merging with master
-  async applyAction() {
-    const resource = this.resources[0];
-    const appPartsData = resource?.applicationParts.reduce((accumulator, currentValue) => {
-      accumulator[currentValue] = {};
-
-      return accumulator;
-    }, {});
-
-    const chartZip = async(files) => {
-      const zip = new JSZip();
-
-      for (const fileName in files) {
-        const extension = {
-          [APPLICATION_PARTS.VALUES]: 'yml',
-          [APPLICATION_PARTS.CHART]:  'tar.gz',
-          [APPLICATION_PARTS.IMAGE]:  'tar',
-        };
-
-        zip.file(`${ fileName }.${ extension[fileName] }`, files[fileName]);
-      }
-
-      const contents = await zip.generateAsync({ type: 'blob' });
-
-      await downloadFile(`${ resource.meta.name }-helm-chart.zip`, contents, 'application/zip');
-    };
-
-    if (this.$route.hash === '#manifest') {
-      await resource.createManifest();
-    } else {
-      await Promise.all(resource?.applicationParts.map(async(part) => {
-        const data = await resource.fetchPart(part);
-
-        appPartsData[part] = data;
-      }));
-
-      await chartZip(appPartsData);
     }
   }
